+++ internal/domain/database_networking.go
@@ -1,0 +1,12 @@
+package domain
+
+type DatabaseNetworkConfiguration struct {
+	OutscaleAccountID string
+	OutscaleNetID     string
+	IPRange           string
+}
+
+type DatabaseNetPeering struct {
+	ID                   string
+	OutscaleNetPeeringID string
+}